*: update all lookup_entry() callbacks
This is a follow-up to the previous commit. Using the
nb_oper_data_iterate() API function it's now possible to specify a
base offset of where the iteration should start. Since a northbound
client can perform this iteration asynchronously, retrieving data
in small chunks, it's possible that an iteration might stop in the
middle of a list entry. When the iteration is resumed, this list
entry might not exist anymore. In this case, the lookup_entry()
callback would fail and consequently the data iteration would be
aborted. This shouldn't be handled as an error, however, since data
is allowed to change while it's been iterated over.

To address this issue, introduce a new "exact_match" (boolean)
parameter to the lookup_entry() callback. This parameter will be
set to 'false' only when starting a data iteration from a base
offset, otherwise it's set to 'true'. When it's set to 'false',
the northbound callbacks should return the first list entry greater
than or equal to the provided search keys. This means that, if
the requested list entry was removed, the next one in the list is
returned instead and the iteration continues.

Here's a summary of how this can be done:
* When iterating over rb-trees, use RB_FIND for exact matches and
  RB_NFIND for non-exact matches;
* When iterating over routing tables, use route_node_lookup()
  for exact matches and route_table_get_next() for non-exact matches;
* For linked-lists, use == or >= logic for exact and non-exact
  matches when comparing elements (this assumes the list is ordered,
  otherwise it isn't an appropriate data-structure for storing
  YANG-modeled state data);
* For hash tables there's no easy way to perform a lookup with >=
  semantics, since hash tables store elements out of order by
  definition. In this case, the best thing to do is to switch to
  a different data structure that allow efficient in-order traversal
  (e.g. rb-trees). The lookup_entry() callbacks from bfdd and vrrpd
  fall into this category and weren't updated for this reason.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>