lib: extend the Get() gRPC RPC to support operational-data chunking
The Get() RPC was extended with two new input parameters ("offset"
and "chunk_size") and one new output parameter ("offset") in order
to leverage the new data chunking capabilities provided by the
nb_oper_data_iterate() API.

The way it's supposed to be used is very simple: if data
fragmentation is desired, Get() should be called inside of a loop
where the input offset is taken from the outputted offset of the
previous call (except in the case of the first call, where no start
offset should be given). Once Get() doesn't return any offset,
it means all data was already retrieved.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>